Modify MonkeyPatching of Phases
The monkey patch of `Phases` to support the appending of interfaces was referencing the wrong append fn, hence the proper interfacing was not established.
Modification of the sample, now when you do `obj.phases = phase` the phase is appended and `obj.phases = new_phases` overwrites the `obj.phases` field. sw